We wrote and tested PipeOutputStream

+++ libs/iostream/vsrc/CompositeOutputStream.vala
@@ -48,6 +48,7 @@ public class shotodol.CompositeOutputStream : OutputStream {
 	}
 
 	public override void close() throws IOStreamError.OutputStreamError {
+		if(closed)return;
 		closed = true;
 		collection.destroy();
 	}
